requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bunch==1.0.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-generic-filters==1.10
django-import-export==2.0.2
django-js-asset==1.2.2
django-user-agents==0.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
ua-parser==0.10.0
user-agents==2.1
xlrd==1.2.0
xlwt==1.3.0
